requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
comtypes==1.4.8
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-speech==2.28.1
google-cloud-texttospeech==2.21.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
idna==3.10
lxml==5.3.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydub==0.25.1
pypiwin32==223
python-docx==1.1.2
pyttsx3==2.98
pywin32==308
requests==2.32.3
rsa==4.9
SpeechRecognition==3.11.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
